+++ views/config/reset-all-to-default.php
@@ -6,7 +6,7 @@
  * @author [Marc FARRE](https://marc.fun) for [CUZY.APP](https://www.cuzy.app)
  */
 
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\cleanTheme\models\Configuration;
 use humhub\modules\ui\view\components\View;
 use yii\helpers\Url;