+++ docs/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased (wait for DeviceDetectorHelper::getBodyClasses() to be available)
 --------------------
+- Cgh: Minimal HumHub version is now 1.18
+- Chg: SCSS instead of LESS
+- Chg: Remove the modification of the HumHub SCSS colors variables (we assume all colors are now CSS variables) 
 - Enh: Add body classes from the new `DeviceDetectorHelper` in 1.17
 
 2.2.0 (November 26, 2024)